# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Restaurante CRUD",
        "description": "API para la gestión de Restaurantes: Comensales, Mesas y Reservas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/Mesa/crudMesa": {
            "post": {
                "tags": [
                    "Mesa"
                ],
                "summary": "Crear, editar o eliminar una mesa",
                "description": "Dependiendo del valor del parámetro 'action', se puede crear (c), actualizar (u) o eliminar (d) una mesa.",
                "operationId": "eb5c0f0b5aad32738423efa8b90a0224",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "description": "Acción a realizar: 'c' para crear, 'u' para actualizar, 'd' para eliminar",
                                        "type": "string",
                                        "example": "c"
                                    },
                                    "id": {
                                        "description": "ID de la mesa (requerido para 'u' y 'd')",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "numero_mesa": {
                                        "description": "Número de la mesa (requerido para 'c' y 'u')",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "capacidad": {
                                        "description": "Capacidad de la mesa (requerido para 'c' y 'u')",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "ubicacion": {
                                        "description": "Ubicación de la mesa (opcional para 'c' y 'u')",
                                        "type": "string",
                                        "example": "ZONA VIP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operación exitosa"
                                        },
                                        "data": {
                                            "properties": {
                                                "mesa": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "numero_mesa": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "capacidad": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "ubicacion": {
                                                            "type": "string",
                                                            "example": "ZONA VIP"
                                                        },
                                                        "state": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2024-04-21T14:22:00.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2024-04-21T14:22:00.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        },
                                        "data": {
                                            "properties": {
                                                "error_message": {
                                                    "type": "string",
                                                    "example": "Detalle del error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}